Move stories to respective component dir
