+++ src/components/Button/Button.stories.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
 
-import { Input } from '../components/Input';
+import { Button } from '.';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-	title: 'Components/Input',
-	component: Input,
+	title: 'Components/Button',
+	component: Button,
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template = (args) => <Input {...args} />;
+const Template = (args) => <Button {...args} />;
 
 export const Preview = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 Preview.args = {
-	value: 'aekstr0m',
-	placeholder: 'Insert username',
+	primary: true,
+	children: 'Button',
 };
