+++ src/components/Input/Input.stories.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
 
-import { Button } from '../components/Button';
+import { Input } from '.';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-	title: 'Components/Button',
-	component: Button,
+	title: 'Components/Input',
+	component: Input,
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template = (args) => <Button {...args} />;
+const Template = (args) => <Input {...args} />;
 
 export const Preview = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 Preview.args = {
-	primary: true,
-	children: 'Button',
+	value: 'aekstr0m',
+	placeholder: 'Insert username',
 };
